requirements.txt
flask
gunicorn
langchain-core
langchain-google-genai
openpyxl
pandas
python-dotenv
requests
twilio